guest_memory: Fix region access violation
While call try_access from read_from, in try_access, it does not
check if start + len will over the end of current region
before it pass the region to the closure, in which, the region
will be converted to slice, and data is picked up from the slice,
from the start to the end = start + len, those parameters are
all transferred from try_access. If the end is bigger than
the end of the region, then region access violation will happen.

Get the len in a differen way to ensure it will not over the region
from try_access to avoid such error.

Signed-off-by: Cathy Zhang <cathy.zhang@intel.com>